Moving sponsors to all pages.
